// client/src/app/types/enums/ingredient.enum.ts
export enum Ingredient {
    TOMATO_SAUCE = 'TOMATO_SAUCE',
    MOZZARELLA = 'MOZZARELLA',
    HAM = 'HAM',
    OREGANO = 'OREGANO',
    GARLIC = 'GARLIC',
    ONION = 'ONION',
    MUSHROOMS = 'MUSHROOMS',
    PARMESAN = 'PARMESAN',
    BASIL = 'BASIL',
    PEPPERONI = 'PEPPERONI',
    EGG = 'EGG',
    CHILLI_PEPPER = 'CHILLI_PEPPER',
    SOUR_CREAM = 'SOUR_CREAM',
    BACON = 'BACON',
    BLUE_CHEESE = 'BLUE_CHEESE',
    GORGONZOLA = 'GORGONZOLA',
    TUNA = 'TUNA',
    OLIVES = 'OLIVES',
}